gpu: nvgpu: fix static violations in common
 * Updated types and added error checks
 * Modified GR condition for ctxsw disable count
   CERT-C error check was added to detect error on integer overflow
   But below logic couldn't detect first overflow, so updated condition

   INT_MAX < gr->ctxsw_disable_count --> it became true after overflow
   So, we didn't detected in first overflow and lead to assert on enable

JIRA NVGPU-3400

Reviewed-on: https://git-master.nvidia.com/r/2206282
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/common/acr/acr_priv.h
@@ -115,7 +115,7 @@ struct nvgpu_acr {
 	u32 bootstrap_owner;
 
 	/* LSF properties */
-	u32 lsf_enable_mask;
+	u64 lsf_enable_mask;
 	struct acr_lsf_config lsf[FALCON_ID_END];
 
 	/*
